feat:이미지텍스트 추출에 따른 Alarm 도메인의 MVC를 재구성

+++ src/main/java/com/kgu/bravoHealthPark/domain/alarm/controller/AlarmController.java
@@ -2,22 +2,21 @@ package com.kgu.bravoHealthPark.domain.alarm.controller;
 
 import com.kgu.bravoHealthPark.domain.alarm.domain.Alarm;
 import com.kgu.bravoHealthPark.domain.alarm.domain.AlarmStatus;
+import com.kgu.bravoHealthPark.domain.alarm.domain.Meal;
 import com.kgu.bravoHealthPark.domain.alarm.dto.AlarmDto;
 import com.kgu.bravoHealthPark.domain.alarm.dto.AlarmForm;
 import com.kgu.bravoHealthPark.domain.alarm.service.AlarmService;
 import com.kgu.bravoHealthPark.domain.medicationInfo.domain.MedicationInfo;
 import com.kgu.bravoHealthPark.domain.medicationInfo.service.MedicationInfoService;
-import com.kgu.bravoHealthPark.domain.user.domain.User;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -31,33 +30,51 @@ public class AlarmController {
 
     @ApiOperation(value = "알람 생성")
     @PostMapping("/create")
-    public ResponseEntity<AlarmDto> createAlarm(Long medicationInfoId, String title, String time) {
+    public ResponseEntity<List<AlarmDto>> createAlarm(Long medicationInfoId, Meal meal, String... times) {
         MedicationInfo medicationInfo = medicationInfoService.findByMedicationInfoId(medicationInfoId);
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("HH:mm");
-        LocalTime localTime = LocalTime.parse(time, formatter);
+        int days = medicationInfo.getDays();
 
-        Alarm alarm = new Alarm(medicationInfo, title, localTime);
-        alarm.initStatus();
+        List<AlarmDto> alarmDtoList = new ArrayList<>();
 
-        alarmService.save(alarm);
+        for (String time : times) {
+            DateTimeFormatter formatter = DateTimeFormatter.ofPattern("HH:mm");
+            LocalTime localTime = LocalTime.parse(time, formatter);
 
-        AlarmDto alarmDto = new AlarmDto(alarm);
+            if (meal == Meal.BEFORE_MEAL)
+                localTime = localTime.minusMinutes(30);
+            else
+                localTime = localTime.plusMinutes(30);
 
-        return ResponseEntity.ok().body(alarmDto);
+            Alarm alarm = new Alarm(medicationInfo, "약 먹을 시간입니다", localTime, meal);
+            alarm.initStatus();
+
+            alarmService.save(alarm);
+
+            AlarmDto alarmDto = new AlarmDto(alarm);
+            alarmDtoList.add(alarmDto);
+        }
+
+        return ResponseEntity.ok().body(alarmDtoList);
     }
 
     @ApiOperation(value = "알람 삭제")
     @DeleteMapping("/delete/{alarmId}")
     public ResponseEntity<AlarmDto> deleteAlarm(@PathVariable Long alarmId) {
         Alarm findAlarm = alarmService.findAlarmById(alarmId);
-        alarmService.deleteAlarm(findAlarm);
+
+        // 연관된 객체를 해제
+        MedicationInfo medicationInfo = findAlarm.getMedicationInfo();
+        if (medicationInfo != null) {
+            findAlarm.deleteMedicationInfo();
+            alarmService.deleteAlarm(findAlarm);
+        }
 
         return ResponseEntity.ok().body(null);
     }
 
     @ApiOperation(value = "알람 수정")
     @PatchMapping("/update/{alarmId}")
-    public ResponseEntity<AlarmDto> updateAlarm(@PathVariable Long alarmId, @RequestParam AlarmForm form) {
+    public ResponseEntity<AlarmDto> updateAlarm(@PathVariable Long alarmId, @RequestBody AlarmForm form) {
         Alarm findAlarm = alarmService.findAlarmById(alarmId);
         alarmService.updateAlarm(findAlarm, form);
 
@@ -66,7 +83,7 @@ public class AlarmController {
         return ResponseEntity.ok().body(alarmDto);
     }
 
-    @ApiOperation(value = "알람 확인후 복용 상태 ")
+    @ApiOperation(value = "알람 확인후 복용 상태 변경")
     @PatchMapping("/update/alarmstatus/dose/{alarmId}")
     public ResponseEntity<AlarmDto> updateAlarmStatusDose(@PathVariable Long alarmId) {
         Alarm findAlarm = alarmService.findAlarmById(alarmId);
@@ -77,7 +94,7 @@ public class AlarmController {
         return ResponseEntity.ok().body(alarmDto);
     }
 
-    @ApiOperation(value = "알람 확인후 복용하지 않음 상태")
+    @ApiOperation(value = "알람 확인후 복용하지 않음 상태로 변경")
     @PatchMapping("/update/alarmstatus/notdose/{alarmId}")
     public ResponseEntity<AlarmDto> updateUncheck(@PathVariable Long alarmId) {
         Alarm findAlarm = alarmService.findAlarmById(alarmId);
@@ -98,22 +115,10 @@ public class AlarmController {
         return ResponseEntity.ok().body(result);
     }
 
-    @ApiOperation(value = "알람 전체 찾기")
-    @GetMapping("/all")
-    public ResponseEntity<List<AlarmDto>> searchAllAlarm() {
-        List<Alarm> alarmList = alarmService.findAlarmAll();
-
-        List<AlarmDto> result = alarmList.stream()
-                .map(a -> new AlarmDto(a))
-                .collect(Collectors.toList());
-
-        return ResponseEntity.ok().body(result);
-    }
-
-    @ApiOperation(value = "알람 이름으로 알람 찾기")
-    @GetMapping("/search/title")
-    public ResponseEntity<List<AlarmDto>> searchAlarmByTitle(@RequestParam String title) {
-        List<Alarm> alarmList = alarmService.findAlarmByTitle(title);
+    @ApiOperation(value = "유저별 알람 찾기")
+    @GetMapping("/search/user/{userId}")
+    public ResponseEntity<List<AlarmDto>> searchAlarmByUser(@PathVariable Long userId) {
+        List<Alarm> alarmList = alarmService.findAlarmByUser(userId);
 
         List<AlarmDto> result = alarmList.stream()
                 .map(a -> new AlarmDto(a))
@@ -126,18 +131,6 @@ public class AlarmController {
     @GetMapping("/search/status")
     public ResponseEntity<List<AlarmDto>> searchAlarmByStatus(@RequestParam AlarmStatus alarmStatus) {
         List<Alarm> alarmList = alarmService.findAlarmByStatus(alarmStatus);
-
-        List<AlarmDto> result = alarmList.stream()
-                .map(a -> new AlarmDto(a))
-                .collect(Collectors.toList());
-
-        return ResponseEntity.ok().body(result);
-    }
-
-    @ApiOperation(value = "유저별 알람 찾기")
-    @GetMapping("/search/user/{userId}")
-    public ResponseEntity<List<AlarmDto>> searchAlarmByUser(@PathVariable Long userId) {
-        List<Alarm> alarmList = alarmService.findAlarmByUser(userId);
 
         List<AlarmDto> result = alarmList.stream()
                 .map(a -> new AlarmDto(a))
